adding require auth check to KN controller

+++ API/Controllers/KnowledgeNodeController.cs
@@ -1,12 +1,14 @@
-﻿using System;
+﻿using Azure.Core;
+using Knowledge_Center.Models;
+using Knowledge_Center.Services.Core;
+using Knowledge_Center.Services.Security;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
-using Knowledge_Center.Models;
-using Knowledge_Center.Services.Core;
 
 namespace Knowledge_Center.API.Controllers
 {
@@ -49,6 +51,9 @@ namespace Knowledge_Center.API.Controllers
         // === POST /api/knowledge-nodes ===
         public void Create(HttpListenerResponse response, HttpListenerRequest request)
         {
+            // Checks for Auth first
+            if (!AuthHelper.RequireAuth(request, response)) return; 
+
 
             using var reader = new StreamReader(request.InputStream, request.ContentEncoding);
             string body = reader.ReadToEnd();
@@ -78,6 +83,9 @@ namespace Knowledge_Center.API.Controllers
         // === PUT /api/knowledge-nodes/{id} ===
         public void Update(HttpListenerResponse response, HttpListenerRequest request, int id) 
         {
+            // Checks for Auth first
+            if (!AuthHelper.RequireAuth(request, response)) return;
+
             using var reader = new StreamReader(request.InputStream, request.ContentEncoding);
             string body = reader.ReadToEnd();
 
@@ -105,8 +113,11 @@ namespace Knowledge_Center.API.Controllers
         }
 
         // === DELETE /api/knowledge-nodes/{id} === 
-        public void Delete(HttpListenerResponse response, int id)
+        public void Delete(HttpListenerRequest request, HttpListenerResponse response, int id)
         {
+            // Checks for Auth first
+            if (!AuthHelper.RequireAuth(request, response)) return;
+
             // Delete all logs associated with the Knowledge Node
             bool logsDeleted = _logEntryService.DeleteAllLogEntriesByNodeId(id);
